Move api TypeScript definitios to dedicated api folder
