+++ src/endpoints/poisons/PoisonsView.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useMemo, useState } from 'react';
 import { DataTable, DataTableSearchInput, type ColumnDef } from '../../components/DataTable'
-import { fetchPoisons, upsertPoison, deletePoison } from './api';
+import { fetchPoisons, upsertPoison, deletePoison } from '../../api/poisons';
 import type { Poison } from '../../types';
 import { useConfirm } from '../../components/ConfirmDialog';
 import { useToast } from '../../components/Toast';
